Authentication

+++ src/main/java/com/github/sba/authentication/service/impl/UserServiceImpl.java
@@ -8,18 +8,25 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.github.sba.authentication.bean.UserBean;
 import com.github.sba.authentication.entity.QUser;
 import com.github.sba.authentication.entity.User;
+import com.github.sba.authentication.exceptions.PermissionDeniedException;
 import com.github.sba.authentication.exceptions.UserAlreadyExistsException;
 import com.github.sba.authentication.exceptions.UserInvalidFieldException;
 import com.github.sba.authentication.exceptions.UserNotFoundException;
 import com.github.sba.authentication.exceptions.UserNotNullException;
 import com.github.sba.authentication.repository.UserRepository;
+import com.github.sba.authentication.security.UserLoggedService;
 import com.github.sba.authentication.service.UserService;
 import com.querydsl.core.types.dsl.BooleanExpression;
 
 @Service
 public class UserServiceImpl implements UserService {
+
+	@Autowired
+	private UserLoggedService userLoggedService;
+
 	@Autowired
 	private UserRepository userRepository;
 
@@ -30,14 +37,14 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void requiredFields(User user) {
-		if (user == null)
+	public void requiredFields(UserBean userBean) {
+		if (userBean == null)
 			throw new UserNotNullException("Usuário não encontrado");
 
-		if (user.getEmail() == null || user.getEmail().trim().equals(""))
+		if (userBean.getEmail() == null || userBean.getEmail().trim().equals(""))
 			throw new UserInvalidFieldException("Email não preenchido");
 
-		if (user.getName() == null || user.getName().trim().equals(""))
+		if (userBean.getName() == null || userBean.getName().trim().equals(""))
 			throw new UserInvalidFieldException("Nome não preenchido");
 	}
 
@@ -51,19 +58,24 @@ public class UserServiceImpl implements UserService {
 		BooleanExpression expression = qUser.email.eq((String) email).and(qUser.active.isTrue());
 		User user = userRepository.findOne(expression);
 
-		this.requiredFields(user);
+		if (user == null)
+			throw new UserNotNullException("Usuário não encontrado");
 
 		return user;
 	}
 
 	@Transactional
 	@Override
-	public User create(User user) {
-		this.requiredFields(user);
+	public User create(UserBean userBean) {
+		this.requiredFields(userBean);
 
-		if (userRepository.findOne(user.getEmail()) != null)
+		if (userRepository.findOne(userBean.getEmail()) != null)
 			throw new UserAlreadyExistsException("Usuário já cadastrado");
 
+		User user = new User();
+		user.setEmail(userBean.getEmail());
+		user.setPassword(userBean.getPassword());
+		user.setName(userBean.getName());
 		user.setCreateDataTime(new Timestamp(System.currentTimeMillis()));
 		user.setUpdateDateTime(new Timestamp(System.currentTimeMillis()));
 		user.setActive(true);
@@ -72,17 +84,20 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Transactional
-	public User update(User user) {
-		this.requiredFields(user);
+	public User update(UserBean userBean) {
+		this.requiredFields(userBean);
 
-		User userUpdate = userRepository.findOne(user.getEmail());
+		if (!userBean.getEmail().equals(userLoggedService.get().getEmail()))
+			throw new PermissionDeniedException("Permissão negada, você só pode atualizar seus dados");
+
+		User userUpdate = userRepository.findOne(userBean.getEmail());
 
 		if (userUpdate == null)
 			throw new UserNotFoundException("Usuário não encontrado");
 
-		userUpdate.setName(user.getName());
+		userUpdate.setName(userBean.getName());
 		userUpdate.setUpdateDateTime(new Timestamp(System.currentTimeMillis()));
-		userUpdate.setActive(user.isActive());
+		userUpdate.setActive(userBean.isActive());
 
 		return userRepository.save(userUpdate);
 	}
@@ -93,7 +108,12 @@ public class UserServiceImpl implements UserService {
 			throw new UserNotNullException("Email não pode ser nulo");
 
 		User user = userRepository.findOne(email);
-		this.requiredFields(user);
+
+		if (user == null)
+			throw new UserNotNullException("Usuário não encontrado");
+
+		if (!email.equals(userLoggedService.get().getEmail()))
+			throw new PermissionDeniedException("Permissão negada, você só pode excluir seus dados");
 
 		userRepository.delete(user);
